requirements.txt
streamlit==1.37.0
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.5
XlsxWriter==3.2.0
openai>=1.40.0
python-dotenv>=1.0.1
streamlit-authenticator==0.3.2
bcrypt==4.0.1